Move toolbar to the nav controller.
Replace it in UIViewController with `toolbarItems`. This is an `IBOutletCollection(UIBarButtonItem)` property, meaning that one can add any number of `UIBarButtonItem` objects to it right in Interface Builder (requires iOS 4.0 or higher).

The advantage to this is that each contained view only needs to managed its UIBarButtonItems in IB, and when we change views, the nav controller just replaces the toolbar items in the toolbar, without having to do any fancy animation to swap toolbars. The latter is something we would have to figure out, and is probably mostly a waaste. The former is built in to UIToolbar.

Things still to think about:

* How to provide control over the appearance of the toolbar.
* How to specify the type of animation when pushing and popping view controllers.
* An API to replace the toolbar whole hog (and the animation to go with it).